+++ docker-entrypoint.sh
@@ -2,7 +2,7 @@
 set -euo pipefail
 
 echo "[entrypoint] Backfilling today's messages for enabled chats..."
-uv run python src/backfill_all_today.py || echo "[entrypoint] Backfill step finished with non-zero exit (continuing to start the bot)"
+uv run python -m src.backfill.backfill_all_today || echo "[entrypoint] Backfill step finished with non-zero exit (continuing to start the bot)"
 
 echo "[entrypoint] Starting bot..."
-exec uv run python src/main.py
+exec uv run python -m src.main
