+++ src/backfill/backfill_all_today.py
@@ -1,36 +1,17 @@
 import os
 import sys
-import sqlite3
 import subprocess
-from contextlib import closing
 from datetime import datetime
 from zoneinfo import ZoneInfo
 from pathlib import Path
 from dotenv import load_dotenv
 
 load_dotenv()
-DB_PATH = os.getenv("DB_PATH", "bot.db")
-TZ = os.getenv("TZ", "Europe/Kyiv")
 
-SCRIPT_DIR = Path(__file__).resolve().parent
-PROJECT_ROOT = SCRIPT_DIR.parent
-
-def db():
-    conn = sqlite3.connect(DB_PATH)
-    conn.row_factory = sqlite3.Row
-    return conn
-
-def get_enabled_chat_ids():
-    with closing(db()) as conn, closing(conn.cursor()) as cur:
-        cur.execute("SELECT chat_id FROM chats WHERE enabled=1")
-        return [row["chat_id"] for row in cur.fetchall()]
-
-def ensure_schema_exists():
-    # Ensure we can import main from src/ to initialize schema without starting the bot.
-    # main.py guards execution with if __name__ == "__main__": so import is safe.
-    if str(SCRIPT_DIR) not in sys.path:
-        sys.path.insert(0, str(SCRIPT_DIR))
-    import main  # noqa: F401  # triggers schema creation at import time
+# This script is run as a module from the project root,
+# so we can import directly from `src`.
+from src.db import get_enabled_chat_ids, init_db
+from src.config import TZ
 
 def main():
     api_id = os.getenv("TELEGRAM_API_ID")
@@ -39,7 +20,7 @@ def main():
         print("TELEGRAM_API_ID/TELEGRAM_API_HASH not set, skipping backfill.")
         return 0
 
-    ensure_schema_exists()
+    init_db()
 
     chat_ids = get_enabled_chat_ids()
     if not chat_ids:
@@ -51,14 +32,14 @@ def main():
     date_str = today.strftime("%Y-%m-%d")
 
     print(f"Backfilling {date_str} for {len(chat_ids)} enabled chats...")
-    backfill_script = SCRIPT_DIR / "backfill_today.py"
     for cid in chat_ids:
         print(f"  - Chat {cid}: backfill today")
         try:
+            # Run the child script as a module to ensure its sys.path is correct.
             subprocess.run(
                 [
                     sys.executable,
-                    str(backfill_script),
+                    "-m", "src.backfill.backfill_today",
                     "--api-id", str(int(api_id)),
                     "--api-hash", api_hash,
                     "--chat-id", str(cid),
@@ -66,7 +47,6 @@ def main():
                     "--date", date_str,
                 ],
                 check=True,
-                cwd=str(PROJECT_ROOT),
             )
         except subprocess.CalledProcessError as e:
             print(f"    Backfill failed for chat {cid}: {e}. Continuing...")
